Harden parsers against corrupt/adversarial data (broad review)
From the whole-crate review. The security-critical surface (crypto, manifest,
path handling, SQL) verified clean; these are the data-integrity fixes:

- Timestamp overflow (13 sites): `d as i64 + MAC_EPOCH` saturates the f64→i64
  cast to i64::MAX and then overflows the i64 add — a panic in dev builds
  (overflow-checks on) or a wrapped negative time in release, on a corrupt date
  (~1e19). Now adds in f64 before the cast so it saturates cleanly, matching the
  already-correct notes/recordings/camera_roll sites. Covers safari, calls,
  address_book, photos_meta, reminders, health, coreduet, calendar, and the
  viber/whatsapp/kik/threema app parsers. Regression test on the shared pattern.
- Safari bookmarks: a NULL `type` or `id` was read as a required i64, so one
  such row aborted the entire bookmarks/reading-list/tabs load. NULL `type` → not
  a folder; NULL `id` → skip the unkeyable row.
- WhatsApp / Facebook Messenger: `body` (and FB's `timestamp`) read strictly, so
  one BLOB-typed cell discarded that app's whole message set. Now use the tolerant
  col_string/col_i64 helpers the sibling app parsers already use.
- Recordings re-import: the INSERT hardcoded `folder = NULL`, dropping the
  Voice-Memos folder name on a recordings-only re-import. Now stores rec.folder,
  matching the full-import path.

cargo test (105) / clippy -D warnings / fmt clean.

+++ crates/traceloupe-core/src/parsers/address_book.rs
@@ -76,7 +76,7 @@ pub fn parse_address_book(db_path: &Path) -> Result<Vec<ParsedContact>> {
         let birthday_at = r
             .get::<_, Option<String>>(11)?
             .and_then(|s| s.trim().parse::<f64>().ok())
-            .map(|t| t as i64 + MAC_EPOCH);
+            .map(|t| (t + MAC_EPOCH as f64) as i64);
         Ok(Row {
             id: r.get(0)?,
             first: r.get(1)?,
