+++ crates/traceloupe-core/src/parsers/apps/threema.rs
@@ -97,7 +97,7 @@ fn parse(db_path: &Path, _rel_path: &str) -> Result<Vec<AppMessage>> {
         let timestamp = r
             .get::<_, Option<f64>>(4)?
             .filter(|d| *d > 0.0)
-            .map(|d| d as i64 + MAC_EPOCH);
+            .map(|d| (d + MAC_EPOCH as f64) as i64);
         // ZSENDER present ⇒ an explicit per-message author (a group message).
         let sender_pk: Option<i64> = super::col_i64(r, 5)?;
         let sender_name: Option<String> = super::col_string(r, 6)?.filter(|s| !s.trim().is_empty());
